WIP: a /v1/user_login endpoint
This accepts a POST call and asks condor for a token with your authenticated
username.  You can send it some JSON with a "claimtobe" field to ask to be
mapped as the claimtobe user, or send "mock: true" to get a fake token.
(I'm not sure I need both of those.)

I didn't use much of flask_restful because I'm still planning on getting rid
of it eventually.